+++ lib/flutter_litert.dart
@@ -17,6 +17,8 @@ export 'src/util/image_tensor_utils.dart';
 export 'src/util/letterbox_params.dart';
 export 'src/util/model_output_utils.dart';
 export 'src/util/yuv_conversion.dart';
+export 'src/util/camera_frame.dart';
+export 'src/util/camera_overlay.dart';
 export 'src/performance_config.dart';
 export 'src/ssd_anchors.dart';
 export 'src/round_robin_pool.dart';